peripheral: fix [XJ3-4097] Fix warning about implicit type conversion from wide to narrow
Summary:
fix warning for spi,i2c,uart module warning

Test Plan:
pass
drivers/spi/Makefile
drivers/i2c/busses/Makefile
drivers/tty/serial/Makefile
ccflags-y += -Wconversion -Wno-sign-conversion

Reviewers: yunqian.wang, ming.yu

Reviewed By: yunqian.wang

Subscribers: siyuan.wang

Differential Revision: https://cr.hobot.cc/D164429

+++ drivers/spi/spi-hobot-slave.h
@@ -128,12 +128,12 @@ struct spidev_data {
 	int level;
 	struct spi_statistic spi_statistic;
 
-	u32 timeout;
+	u64 timeout;
 
 	u8 *tx_dummy_buffer;
 
 	u8 noblock;
-	u32 data_len;
+	size_t data_len;
 	u8 *tx_swap_buffer;
 
 	struct rx_tmp_buffer spi_recv_buf;
